authorisation/pyproject.toml
[tool.poetry]
name = "authorisation"
version = "0.1.0"
description = ""
authors = ["lukerichardson-dev <richardsonluke232@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.31"
grpcio = "^1.44.0"
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
grpcio-tools = "^1.44.0"
cryptography = "^36.0.1"
psycopg2-binary = { version="^2.9.3", optional = false }
# google = "^3.0.0"
protobuf = "^3.19.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy-protobuf = "^3.2.0"

[tool.poetry.extras]
pgsql = ["psycopg2"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run-service = "authorisation.entrypoint:main"
